+++ egui/src/util/id_type_map.rs
@@ -311,19 +311,19 @@ use crate::Id;
 /// map.insert_temp(a, 42);
 ///
 /// // `b` associated with an f64 and a `&'static str`
-/// map.insert_persisted(b, 6.28);
+/// map.insert_persisted(b, 13.37);
 /// map.insert_temp(b, "Hello World".to_string());
 ///
 /// // we can retrieve all four values:
 /// assert_eq!(map.get_temp::<f64>(a), Some(3.14));
 /// assert_eq!(map.get_temp::<i32>(a), Some(42));
-/// assert_eq!(map.get_temp::<f64>(b), Some(6.28));
+/// assert_eq!(map.get_temp::<f64>(b), Some(13.37));
 /// assert_eq!(map.get_temp::<String>(b), Some("Hello World".to_string()));
 ///
 /// // we can retrieve them like so also:
 /// assert_eq!(map.get_persisted::<f64>(a), Some(3.14));
 /// assert_eq!(map.get_persisted::<i32>(a), Some(42));
-/// assert_eq!(map.get_persisted::<f64>(b), Some(6.28));
+/// assert_eq!(map.get_persisted::<f64>(b), Some(13.37));
 /// assert_eq!(map.get_temp::<String>(b), Some("Hello World".to_string()));
 /// ```
 #[derive(Clone, Debug, Default)]
@@ -544,11 +544,11 @@ fn test_two_id_two_type() {
     let b = Id::new("b");
 
     let mut map: IdTypeMap = Default::default();
-    map.insert_persisted(a, 6.28);
+    map.insert_persisted(a, 13.37);
     map.insert_temp(b, 42);
-    assert_eq!(map.get_persisted::<f64>(a), Some(6.28));
+    assert_eq!(map.get_persisted::<f64>(a), Some(13.37));
     assert_eq!(map.get_persisted::<i32>(b), Some(42));
-    assert_eq!(map.get_temp::<f64>(a), Some(6.28));
+    assert_eq!(map.get_temp::<f64>(a), Some(13.37));
     assert_eq!(map.get_temp::<i32>(b), Some(42));
 }
 
@@ -565,19 +565,19 @@ fn test_two_id_x_two_types() {
     map.insert_temp(a, 42);
 
     // `b` associated with an f64 and a `&'static str`
-    map.insert_persisted(b, 6.28);
+    map.insert_persisted(b, 13.37);
     map.insert_temp(b, "Hello World".to_string());
 
     // we can retrieve all four values:
     assert_eq!(map.get_temp::<f64>(a), Some(3.14));
     assert_eq!(map.get_temp::<i32>(a), Some(42));
-    assert_eq!(map.get_temp::<f64>(b), Some(6.28));
+    assert_eq!(map.get_temp::<f64>(b), Some(13.37));
     assert_eq!(map.get_temp::<String>(b), Some("Hello World".to_string()));
 
     // we can retrieve them like so also:
     assert_eq!(map.get_persisted::<f64>(a), Some(3.14));
     assert_eq!(map.get_persisted::<i32>(a), Some(42));
-    assert_eq!(map.get_persisted::<f64>(b), Some(6.28));
+    assert_eq!(map.get_persisted::<f64>(b), Some(13.37));
     assert_eq!(map.get_temp::<String>(b), Some("Hello World".to_string()));
 }
 
@@ -586,11 +586,11 @@ fn test_one_id_two_types() {
     let id = Id::new("a");
 
     let mut map: IdTypeMap = Default::default();
-    map.insert_persisted(id, 6.28);
+    map.insert_persisted(id, 13.37);
     map.insert_temp(id, 42);
 
-    assert_eq!(map.get_temp::<f64>(id), Some(6.28));
-    assert_eq!(map.get_persisted::<f64>(id), Some(6.28));
+    assert_eq!(map.get_temp::<f64>(id), Some(13.37));
+    assert_eq!(map.get_persisted::<f64>(id), Some(13.37));
     assert_eq!(map.get_temp::<i32>(id), Some(42));
 
     // ------------
@@ -601,8 +601,8 @@ fn test_one_id_two_types() {
     assert_eq!(map.get_temp::<i32>(id), None);
 
     // Other type is still there, even though it is the same if:
-    assert_eq!(map.get_temp::<f64>(id), Some(6.28));
-    assert_eq!(map.get_persisted::<f64>(id), Some(6.28));
+    assert_eq!(map.get_temp::<f64>(id), Some(13.37));
+    assert_eq!(map.get_persisted::<f64>(id), Some(13.37));
 
     // But we can still remove the last:
     map.remove::<f64>(id);